refactor: Home 컴포넌트에서 이미지 정렬 로직 분리
